+++ R/fstg02.R
@@ -1,6 +1,6 @@
-# fstg01 ----
+# fstg02 ----
 
-#' @describeIn fstg01 Main TLG Function
+#' @describeIn fstg02 Main TLG Function
 #'
 #' @details
 #'  * No overall value.
@@ -9,10 +9,9 @@
 #' @inheritParams gen_args
 #' @param dataset (`string`) the name of a table in the `adam_db` object.
 #' @param arm_var (`string`) the arm variable name used for group splitting.
-#' @param rsp_var (`string`) the response variable name to flag whether each subject is a binary response or not.
 #' @param subgroups (`character`) the subgroups variable name to list baseline risk factors.
 #' @param strata_var (`character`) required if stratified analysis is performed.
-#' @param stat_var (`character`) the names of statistics to be reported in `tabulate_rsp_subgroups`.
+#' @param stat_var (`character`) the names of statistics to be reported in `tabulate_survival_subgroups`.
 #' @param ... Further arguments passed to `g_forest` and `extract_rsp_subgroups` (a wrapper for
 #' `h_odds_ratio_subgroups_df` and `h_proportion_subgroups_df`). For details, see the documentation in `tern`.
 #' Commonly used arguments include `col_symbol_size`, `col`, `vline`, `groups_lists`, `conf_level`,
@@ -20,53 +19,56 @@
 #'
 #' @note
 #'  * `adam_db` object must contain the table specified by `dataset` with `"PARAMCD"`, `"ARM"`,
-#'  `"AVALC"`, and the columns specified by `subgroups` which is denoted as
+#'  `"AVAL"`, `"AVALU"`, `"CNSR"`, and the columns specified by `subgroups` which is denoted as
 #'  `c("SEX", "AGEGR1", "RACE")` by default.
 #'
 #' @return a `gTree` object.
 #' @export
-fstg01_main <- function(adam_db,
-                        dataset = "adrs",
+fstg02_main <- function(adam_db,
+                        dataset = "adtte",
                         arm_var = "ARM",
-                        rsp_var = "IS_RSP",
                         subgroups = c("SEX", "AGEGR1", "RACE"),
                         strata_var = NULL,
-                        stat_var = c("n_tot", "n", "n_rsp", "prop", "or", "ci"),
+                        stat_var = c("n_tot", "n", "median", "hr", "ci"),
                         ...) {
   assert_all_tablenames(adam_db, c("adsl", dataset))
   df_lbl <- paste0("adam_db$", dataset)
   assert_string(arm_var)
-  assert_string(rsp_var)
   assert_character(subgroups, null.ok = TRUE)
   assert_character(strata_var, null.ok = TRUE)
   assert_character(stat_var, null.ok = TRUE)
   assert_valid_variable(adam_db[[dataset]], arm_var, types = list("factor"), n.levels = 2, label = df_lbl)
-  assert_valid_variable(adam_db[[dataset]], c("USUBJID", "PARAMCD"),
+  assert_valid_variable(adam_db[[dataset]], c("USUBJID", "PARAMCD", "AVALU"),
     types = list(c("character", "factor")),
     label = df_lbl
   )
-  assert_valid_variable(adam_db[[dataset]], rsp_var, types = list("logical"), label = df_lbl)
+  assert_valid_variable(adam_db[[dataset]], "AVAL", types = list("numeric"), lower = 0, label = df_lbl)
+  assert_valid_variable(adam_db[[dataset]], "IS_EVENT", types = list("logical"), label = df_lbl)
   assert_valid_variable(adam_db[[dataset]], c(subgroups, strata_var),
     types = list(c("factor")), na_ok = TRUE,
     label = df_lbl
   )
   assert_single_value(adam_db[[dataset]]$PARAMCD, label = df_lbl)
+  assert_single_value(adam_db[[dataset]]$AVALU, label = df_lbl)
+
+  timeunit <- unique(adam_db[[dataset]]$AVALU)
 
   variables <- list(
     arm = arm_var,
-    rsp = rsp_var,
+    tte = "AVAL",
+    is_event = "IS_EVENT",
     subgroups = subgroups,
     strata_var = strata_var
   )
 
-  df <- execute_with_args(extract_rsp_subgroups,
+  df <- execute_with_args(extract_survival_subgroups,
     variables = variables,
     data = adam_db[[dataset]],
     ...
   )
 
   result <- basic_table() %>%
-    tabulate_rsp_subgroups(df, vars = stat_var)
+    tabulate_survival_subgroups(df, vars = stat_var, time_unit = timeunit)
 
   execute_with_args(
     g_forest,
@@ -75,26 +77,27 @@ fstg01_main <- function(adam_db,
   )
 }
 
-#' @describeIn fstg01 Preprocessing
+#' @describeIn fstg02 Preprocessing
 #'
-#' @inheritParams fstg01_main
+#' @inheritParams fstg02_main
 #'
 #' @export
-fstg01_pre <- function(adam_db, ...) {
-  adam_db$adrs <- adam_db$adrs %>%
+fstg02_pre <- function(adam_db, ...) {
+  adam_db$adtte <- adam_db$adtte %>%
     mutate(
       ARM = droplevels(.data$ARM),
-      IS_RSP = .data$AVALC %in% c("CR", "PR")
+      AVAL = convert_to_month(.data$AVAL, .data$AVALU),
+      AVALU = "MONTHS",
+      IS_EVENT = .data$CNSR == 0
     )
-
   adam_db
 }
 
-# `fstg01` Pipeline ----
+# `fstg02` Pipeline ----
 
-#' `FSTG01` Subgroup Analysis of Best Overall Response.
+#' `fstg02` Subgroup Analysis of Survival Duration.
 #'
-#' The template produces the subgroup analysis of best overall response graphic.
+#' The template produces the subgroup analysis of survival duration graphic.
 #'
 #' @include chevron_tlg-S4class.R
 #' @export
@@ -105,13 +108,13 @@ fstg01_pre <- function(adam_db, ...) {
 #'
 #' proc_data <- log_filter(
 #'   syn_data,
-#'   PARAMCD == "BESRSPI" & ARM %in% c("A: Drug X", "B: Placebo"), "adrs"
+#'   PARAMCD == "OS" & ARM %in% c("A: Drug X", "B: Placebo"), "adtte"
 #' )
-#' run(fstg01, proc_data,
+#' run(fstg02, proc_data,
 #'   subgroups = c("SEX", "AGEGR1", "RACE"),
-#'   conf_level = 0.90, dataset = "adrs"
+#'   conf_level = 0.90, dataset = "adtte"
 #' )
-fstg01 <- chevron_g(
-  main = fstg01_main,
-  preproces = fstg01_pre
+fstg02 <- chevron_g(
+  main = fstg02_main,
+  preproces = fstg02_pre
 )