+++ NEWS.md
@@ -1,9 +1,12 @@
-# chevron 0.2.1.9001
+# chevron 0.2.1.9002
 
 * Allow `EGT03` to include multiple parameters.
 * Allow `KMG01` to include stratified variables.
 * Allow `LBT06` and `LBT14` split by page.
 * Allow `AET02`, `CMT01A` to change the summary statistics with `summary_labels` argument.
+* Add `FSTG02` template.
+* Update the argument `"is_event"` in `KMG01` to `"IS_EVENT"`.
+* Update the argument `"is_rsp"` in `FSTG01` to `"IS_RSP"`.
 
 # chevron 0.2.1